requirements.txt
aniso8601==9.0.1
cffi==1.15.1
click==8.0.4
cryptography==40.0.2
dataclasses==0.8
Flask==2.0.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
psycopg2-binary==2.9.8
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
